# server/openapi.yaml
openapi: 3.0.3
info:
  title: BIS Platform API
  description: |
    **Background Intelligence System (BIS)** — Nigeria's leading compliance intelligence platform.

    Run KYC checks, sanctions screening, adverse media scans, risk scoring, field agent dispatch,
    and investigation management via a fully metered REST API.

    ## Authentication
    All endpoints require a Bearer token obtained from the BIS Developer Portal (`/developer`).

    ```http
    Authorization: Bearer bis_live_xxxxxxxxxxxxxxxx
    ```

    ## Rate Limits
    Rate limits are enforced per token based on your plan tier:
    - **Starter:** 100 req/min, 10,000 req/day
    - **Professional:** 500 req/min, 100,000 req/day
    - **Enterprise:** Unlimited (fair use)

    ## Token Consumption
    Each API call consumes tokens from your balance. Token costs are shown per endpoint.

    ## Versioning
    This is v1 of the BIS API. Breaking changes will be introduced as v2 with a 6-month deprecation window.
  version: "1.0.0"
  contact:
    name: BIS API Support
    email: api-support@bis-platform.com
    url: https://bis-platform.ng/developer
  license:
    name: Proprietary
    url: https://bis-platform.ng/terms
servers:
  - url: https://bis-platform.ng/api/v1
    description: Production
  - url: https://3000-ixhq6tgivnt84v2lhow3e-f2ebad08.us1.manus.computer/api/v1
    description: Live Demo (read-only)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BIS API Token
...
security:
  - BearerAuth: []
paths:
...
  /openclaw/execute:
    post:
      tags: [OpenClaw]
      summary: Execute OpenClaw action
      description: |
        Managed instance endpoint for the BIS OpenClaw Skill. Executes a named action
        with the given prompt and returns a markdown-formatted result.

        This endpoint is used by the OpenClaw ClawHub skill runner when BIS is configured
        as a managed instance rather than a local skill installation.

        **Tokens: varies by action** (see skill.json for per-action costs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, prompt]
              properties:
                action:
                  type: string
                  enum: [kyc_verify, sanctions_screen, adverse_media, risk_score, create_investigation, dispatch_field_agent, get_investigation, list_alerts, full_due_diligence]
                  example: full_due_diligence
                prompt:
                  type: string
                  example: Run full due diligence on Biodun Afolabi before we sign the JV agreement
                context:
                  type: object
                  nullable: true
                  description: Optional context from the OpenClaw conversation
      responses:
        "200":
          description: Action result as markdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: Markdown-formatted result
                  tokens_consumed:
                    type: integer
                  action:
                    type: string
  /openclaw/webhook:
    post:
      tags: [OpenClaw]
      summary: OpenClaw webhook receiver
      description: |
        Receives webhook events from OpenClaw when an agent completes a BIS-related task.
        Used for async workflows where OpenClaw triggers a BIS investigation and needs
        to be notified when results are ready.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event:
                  type: string
                  enum: [investigation.completed, kyc.completed, field_task.completed, alert.triggered]
                data:
                  type: object
                timestamp:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean